maked it to editable pip package

+++ src/DEXTR/demo.py
@@ -46,7 +46,7 @@ results = []
 
 with torch.no_grad():
     while 1:
-        extreme_points_ori = np.array(plt.ginput(4, timeout=0)).astype(np.int)
+        extreme_points_ori = np.array(plt.ginput(4, timeout=0)).astype(np.int32)
         if extreme_points_ori.shape[0] < 4:
             if len(results) > 0:
                 helpers.save_mask(results, 'demo.png')
@@ -63,7 +63,7 @@ with torch.no_grad():
         #  Generate extreme point heat map normalized to image values
         extreme_points = extreme_points_ori - [np.min(extreme_points_ori[:, 0]), np.min(extreme_points_ori[:, 1])] + [pad,
                                                                                                                       pad]
-        extreme_points = (512 * extreme_points * [1 / crop_image.shape[1], 1 / crop_image.shape[0]]).astype(np.int)
+        extreme_points = (512 * extreme_points * [1 / crop_image.shape[1], 1 / crop_image.shape[0]]).astype(np.int32)
         extreme_heatmap = helpers.make_gt(resize_image, extreme_points, sigma=10)
         extreme_heatmap = helpers.cstm_normalize(extreme_heatmap, 255)
